Upgrade to latest substrate
* Refactored CLI module
* Added IPCI chain genesis to subkey
