+++ bin/node/cli/src/lib.rs
@@ -17,7 +17,6 @@
 ///////////////////////////////////////////////////////////////////////////////
 //! Console line interface.
 
-mod token_holders;
 pub mod chain_spec;
 
 #[macro_use]
@@ -26,8 +25,54 @@ mod service;
 mod browser;
 #[cfg(feature = "cli")]
 mod cli;
+#[cfg(feature = "cli")]
+mod command;
 
 #[cfg(feature = "browser")]
 pub use browser::*;
 #[cfg(feature = "cli")]
 pub use cli::*;
+#[cfg(feature = "cli")]
+pub use command::*;
+
+/// The chain specification option.
+#[derive(Clone, Debug, PartialEq)]
+pub enum ChainSpec {
+	/// Whatever the current runtime is, with just Alice as an auth.
+	Development,
+	/// Whatever the current runtime is, with simple Alice/Bob auths.
+	LocalTestnet,
+    /// Robonomics public testnet.
+    RobonomicsTestnet,
+    /// IPCI blockchain network.
+    IPCI,
+}
+
+impl ChainSpec {
+    /// Get an actual chain config from one of the alternatives.
+    pub(crate) fn load(self) -> chain_spec::ChainSpec {
+        match self {
+            ChainSpec::Development       => chain_spec::development_testnet_config(),
+            ChainSpec::LocalTestnet      => chain_spec::local_testnet_config(),
+            ChainSpec::RobonomicsTestnet => chain_spec::robonomics_testnet_config(),
+            ChainSpec::IPCI              => chain_spec::ipci_config(),
+        }
+    }
+
+    pub(crate) fn from(s: &str) -> Option<Self> {
+        match s {
+            "dev" => Some(ChainSpec::Development),
+            "local" => Some(ChainSpec::LocalTestnet),
+            "" | "robonomics" => Some(ChainSpec::RobonomicsTestnet),
+            "ipci" => Some(ChainSpec::IPCI),
+            _ => None,
+        }
+    }
+}
+
+pub fn load_spec(id: &str) -> Result<Option<chain_spec::ChainSpec>, String> {
+    Ok(match ChainSpec::from(id) {
+        Some(spec) => Some(spec.load()),
+        None => None,
+    })
+}
