+++ bin/subkey/Cargo.toml
@@ -29,7 +29,7 @@ jsonrpc-core-client = { version = "14.0.3", features = ["http"] }
 hyper = "0.12.35"
 
 [features]
-default = ["robonomics"]
+default = ["ipci"]
 robonomics = ["node-executor/robonomics-runtime"]
 ipci = ["node-executor/ipci-runtime"]
 bench = []
