+++ bin/subkey/src/rpc.rs
@@ -26,24 +26,24 @@ use sp_core::Bytes;
 pub struct RpcClient { url: String }
 
 impl RpcClient {
-	pub fn new(url: String) -> Self { Self { url } }
+    pub fn new(url: String) -> Self { Self { url } }
 
-	pub fn insert_key(
-		&self,
-		key_type: String,
-		suri: String,
-		public: Bytes,
-	) {
-		let url = self.url.clone();
+    pub fn insert_key(
+        &self,
+        key_type: String,
+        suri: String,
+        public: Bytes,
+    ) {
+        let url = self.url.clone();
 
-		rt::run(
-			http::connect(&url)
-				.and_then(|client: AuthorClient<Hash, Hash>| {
-					client.insert_key(key_type, suri, public).map(|_| ())
-				})
-				.map_err(|e| {
-					println!("Error inserting key: {:?}", e);
-				})
-		);
-	}
+        rt::run(
+            http::connect(&url)
+                .and_then(|client: AuthorClient<Hash, Hash>| {
+                    client.insert_key(key_type, suri, public).map(|_| ())
+                })
+                .map_err(|e| {
+                    println!("Error inserting key: {:?}", e);
+                })
+        );
+    }
 }
